Port files and add registration
Being registered doesn't actully mean anything yet, other than now your name is in the db.
Added a json file for map info. Getting it programmatically and storing in the db might be better down the line, but this is good enough for now. I might need to store specific beatmap ids if the beatmap set id isn't returned in scores. If that's the case I'll definitely make a way to gather ids automatically.